backend/requirements.txt
flask
langchain
langchain-community
langchain-openai
openai
faiss-cpu
tiktoken
python-dotenv
